Add mouseWidgetTypes and dragAndDropHandler/types
