+++ src/mouseWidgetTypes.ts
@@ -1,0 +1,6 @@
+export interface PositionInfo {
+    pageX: number | undefined;
+    pageY: number | undefined;
+    target: HTMLElement;
+    originalEvent: Event;
+}
